+++ native_handler.go
@@ -10,12 +10,10 @@ type NativeMiddlewareFunc func(http.Handler) http.Handler
 // NativeHandler is an adapter function for "http.Handler".
 // With this adapter, you can plug non-Goyave handlers to your application.
 //
-// If the request is a JSON request, the native handler will not be able to
-// read the body, as it has already been parsed by the framework and is stored in
-// the "goyave.Request" object. However, form data can be accessed as usual.
-// Just remember that it contains the raw data, which haven't been validated
+// Just remember that the body contains the raw data, which haven't been validated
 // nor converted. This means that native handlers are not guaranteed to work and
-// cannot modify the request data.
+// cannot modify the request data. Request properties, such as headers, can still
+// be modified.
 // Prefer implementing a Goyave handler.
 //
 // This feature is a compatibility layer with the rest of the Golang web ecosystem.