+++ router.go
@@ -87,14 +87,21 @@ func newRouter() *Router {
 		parent:            nil,
 		prefix:            "",
 		hasCORSMiddleware: false,
-		statusHandlers:    make(map[int]Handler, 15),
+		statusHandlers:    make(map[int]Handler, 41),
 		namedRoutes:       make(map[string]*Route, 5),
 		middlewareHolder: middlewareHolder{
 			middleware: make([]Middleware, 0, 3),
 		},
 	}
 	router.StatusHandler(panicStatusHandler, http.StatusInternalServerError)
-	router.StatusHandler(errorStatusHandler, 401, 403, 404, 405, 501, 502, 503, 504, 505, 506, 507, 508, 510, 511)
+	for i := 400; i <= 418; i++ {
+		router.StatusHandler(errorStatusHandler, i)
+	}
+	for i := 421; i <= 426; i++ {
+		router.StatusHandler(errorStatusHandler, i)
+	}
+	router.StatusHandler(errorStatusHandler, 428, 429, 431, 444, 451)
+	router.StatusHandler(errorStatusHandler, 501, 502, 503, 504, 505, 506, 507, 508, 510, 511)
 	router.Middleware(recoveryMiddleware, parseRequestMiddleware, languageMiddleware)
 	return router
 }
@@ -305,7 +312,7 @@ func (r *Router) CORS(options *cors.Options) {
 // Status handlers are inherited as a copy in sub-routers. Modifying a child's status handler
 // will not modify its parent's.
 //
-// Codes in the 500 range and codes 401, 403, 404 and 405 have a default status handler.
+// Codes in the 400 and 500 ranges have a default status handler.
 func (r *Router) StatusHandler(handler Handler, status int, additionalStatuses ...int) {
 	r.statusHandlers[status] = handler
 	for _, s := range additionalStatuses {